moved server config to its own file
